boards/feather-m0-*: fix Arduino feature in Kconfig
With PR #17401 the Arduino feature was introduced for the `feather-m0` base board. Since all derived `feather-m0-*` boards with additional hardware modules simply include `Makefile.features` of the `feather-m0` base board, they also have this feature. However, the `Kconfig` file is defined separately for each of these boards. In PR 17401, it was forgotten to include the Arduino feature in the `Kconfig` file for the derived boards. This was not noticed because the CI did not trigger an error message.